pythonHandler/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.0.3
replicate==0.22.0
gunicorn==21.2.0
pyngrok==7.0.5
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.1.post1
psycopg2-binary==2.9.9